Tidy up examples folder.
