+++ tlsn/examples/discord/README.md
@@ -7,9 +7,11 @@ This involves 3 steps:
 2. Start the (local) notary server
 3. Notarize
 
+P/S: The notary server used in this example is more functional compared to its [simple version](../simple/simple_notary.rs). The simple version is easier to integrate with from prover perspective, whereas this notary server provides additional features like TLS connection with prover, WebSocket endpoint, API endpoints for further customisation etc.
+
 ## Inputs
 
-In this tlsn/examples folder, create a `.env` file.
+In this tlsn/examples/discord folder, create a `.env` file.
 Then in that `.env` file, set the values of the following constants by following the format shown in this [example env file](./.env.example).
 
 | Name          | Example                                                                          | Location                                      |
