+++ tlsn/examples/discord/discord_dm_verifier.rs
@@ -4,7 +4,7 @@ use elliptic_curve::pkcs8::DecodePublicKey;
 
 use tlsn_core::proof::{SessionProof, TlsProof};
 
-/// A simple verifier which reads a proof generated by `simple_prover.rs` from "proof.json", verifies
+/// A simple verifier which reads a proof generated by `discord_dm.rs` from "discord_dm_proof.json", verifies
 /// it and prints the verified data to the console.
 fn main() {
     // Deserialize the proof
@@ -67,13 +67,7 @@ fn main() {
 
 /// Returns a Notary pubkey trusted by this Verifier
 fn notary_pubkey() -> p256::PublicKey {
-    // from ../../../notary-server/fixture/notary/notary.key
-    // converted with `openssl ec -in notary.key -pubout -outform PEM`
+    let pem_file_path = "../../../notary-server/fixture/notary/notary.pub";
 
-    let pem = "-----BEGIN PUBLIC KEY-----
-MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEBv36FI4ZFszJa0DQFJ3wWCXvVLFr
-cRzMG5kaTeHGoSzDu6cFqx3uEWYpFGo6C0EOUgf+mEgbktLrXocv5yHzKg==
------END PUBLIC KEY-----";
-
-    p256::PublicKey::from_public_key_pem(pem).unwrap()
+    p256::PublicKey::read_public_key_pem_file(pem_file_path).unwrap()
 }
