+++ tlsn/examples/simple/README.md
@@ -1,3 +1,5 @@
+# Simple Notarization
+
 This guide will take you through the steps of:
 - starting a `Notary` server
 - running a `Prover` to notarize some web data
@@ -5,7 +7,7 @@ This guide will take you through the steps of:
 
 Note that the TLSNotary protocol assumes that the `Notary` is trusted by the `Verifier`. To minimize the trust, the `Verifier` itself can act as a `Notary`.
 
-# Preliminaries
+## Preliminaries
 
 ### Install rust
 
@@ -14,34 +16,33 @@ If you don't have `rust` installed yet, install it with [rustup](https://rustup.
 curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 ```
 
-# Guide
+## Guide
 Clone this repository first
 
 ```shell
 git clone https://github.com/tlsnotary/tlsn
 ```
 
-### Start a Notary server:
+### Start a simple Notary server:
 
 ```shell
-cd tlsn/notary-server
-cargo run --release
+cd tlsn/tlsn/examples/simple
+cargo run --release --example simple_notary
 ```
 
 The `Notary` server will now be running in the background waiting for connections from a `Prover`. You can switch to another console to run the `Prover`.
 
-For more information on how to configure the `Notary` server, please refer to [this](../../notary-server/README.md#running-the-server).
+P/S: The notary server used in this example is less functional compared to its [advanced version](../../../notary-server). This simple version is easier to integrate with from prover perspective, whereas the advanced version provides additional features like TLS connection with prover, WebSocket endpoint, API endpoints for further customisation etc.
 
 ### Run a simple Prover:
 
 ```shell
-cd tlsn/tlsn/examples
-cargo run --release --example simple_prover
+RUST_LOG=DEBUG,yamux=INFO cargo run --release --example simple_prover
 ```
 
 The notarization session usually takes a few moments and the resulting proof will be written to the "proof.json" file. The proof can then be passed on to the `Verifier` for verification.
 
-The `simple_prover` notarizes <https://example.com> and redacts the `USER_AGENT` HTTP header from the proof for the `Verifier`. You can change the code in `tlsn/tlsn/examples/simple_prover.rs` to meet your needs:
+The `simple_prover` notarizes <https://example.com> and redacts the `USER_AGENT` HTTP header from the proof for the `Verifier`. You can change the code in `tlsn/tlsn/examples/simple/simple_prover.rs` to meet your needs:
 
 - change which server the `Prover` connects to
 - add or remove HTTP request headers
