+++ tlsn/examples/twitter/README.md
@@ -7,9 +7,11 @@ This involves 3 steps:
 2. Start the (local) notary server
 3. Notarize
 
+P/S: The notary server used in this example is more functional compared to its [simple version](../simple/simple_notary.rs). The simple version is easier to integrate with from prover perspective, whereas this notary server provides additional features like TLS connection with prover, WebSocket endpoint, API endpoints for further customisation etc.
+
 ## Inputs
 
-In this tlsn/examples folder, create a `.env` file.
+In this tlsn/examples/twitter folder, create a `.env` file.
 Then in that `.env` file, set the values of the following constants by following the format shown in this [example env file](./.env.example).
 
 | Name            | Example                                                 | Location in Request Headers Section (within Network Tab of Developer Tools)      |
@@ -35,11 +37,11 @@ cargo run --release
 
 The notary server will now be running in the background waiting for connections.
 
-For more information on how to configure the notary server, please refer to [this](../../notary-server/README.md#running-the-server).
+For more information on how to configure the notary server, please refer to [this](../../../notary-server/README.md#running-the-server).
 
 ## Notarize
 
-In this tlsn/examples folder, run the following command:
+In this tlsn/examples/twitter folder, run the following command:
 
 ```sh
 RUST_LOG=debug,yamux=info cargo run --release --example twitter_dm
