+++ tlsn/examples/twitter/twitter_dm.rs
@@ -1,120 +1,165 @@
+/// This example shows how to notarize Twitter DMs.
+///
+/// The example uses the notary server implemented in ../../../notary-server
 use eyre::Result;
 use futures::AsyncWriteExt;
 use hyper::{body::to_bytes, client::conn::Parts, Body, Request, StatusCode};
 use rustls::{Certificate, ClientConfig, RootCertStore};
+use serde::{Deserialize, Serialize};
 use std::{
+    env,
     fs::File as StdFile,
     io::BufReader,
     net::{IpAddr, SocketAddr},
     ops::Range,
     sync::Arc,
 };
-use tlsn_core::proof::TlsProof;
 use tokio::{fs::File, io::AsyncWriteExt as _, net::TcpStream};
-use tokio_rustls::{client::TlsStream, TlsConnector};
+use tokio_rustls::TlsConnector;
 use tokio_util::compat::{FuturesAsyncReadCompatExt, TokioAsyncReadCompatExt};
 use tracing::debug;
 
-use notary_server::{ClientType, NotarizationSessionRequest, NotarizationSessionResponse};
 use tlsn_prover::{Prover, ProverConfig};
 
-const SERVER_DOMAIN: &str = "example.com";
+// Setting of the application server
+const SERVER_DOMAIN: &str = "twitter.com";
+const ROUTE: &str = "i/api/1.1/dm/conversation";
 const USER_AGENT: &str = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36";
 
+// Setting of the notary server — make sure these are the same with those in ../../../notary-server
 const NOTARY_HOST: &str = "127.0.0.1";
 const NOTARY_PORT: u16 = 7047;
-const NOTARY_CA_CERT_PATH: &str = "./rootCA.crt";
-const MAX_TRANSCRIPT_SIZE: usize = 16384;
+const NOTARY_CA_CERT_PATH: &str = "../../../notary-server/fixture/tls/rootCA.crt";
 
-/// Runs a simple Prover which connects to the Notary and notarizes a request/response from
-/// example.com. The Prover then generates a proof and writes it to disk.
-///
-/// Note that the Notary server must be already listening on NOTARY_HOST:NOTARY_PORT
-/// (see README.md "Starting a notary server")
+// Configuration of notarization
+const NOTARY_MAX_TRANSCRIPT_SIZE: usize = 16384;
+
+/// Response object of the /session API
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NotarizationSessionResponse {
+    pub session_id: String,
+}
+
+/// Request object of the /session API
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NotarizationSessionRequest {
+    pub client_type: ClientType,
+    /// Maximum transcript size in bytes
+    pub max_transcript_size: Option<usize>,
+}
+
+/// Types of client that the prover is using
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub enum ClientType {
+    /// Client that has access to the transport layer
+    Tcp,
+    /// Client that cannot directly access transport layer, e.g. browser extension
+    Websocket,
+}
+
 #[tokio::main]
 async fn main() {
-    // Initialize logging
     tracing_subscriber::fmt::init();
 
-    // Establish an encrypted connection with the Notary
-    let (notary_socket, session_id) = connect_to_notary().await;
-    println!("Connected to the Notary");
+    // Load secret variables frome environment for twitter server connection
+    dotenv::dotenv().ok();
+    let conversation_id = env::var("CONVERSATION_ID").unwrap();
+    let client_uuid = env::var("CLIENT_UUID").unwrap();
+    let auth_token = env::var("AUTH_TOKEN").unwrap();
+    let access_token = env::var("ACCESS_TOKEN").unwrap();
+    let csrf_token = env::var("CSRF_TOKEN").unwrap();
 
-    // A Prover configuration using the session_id returned by the Notary
+    let (notary_tls_socket, session_id) = setup_notary_connection().await;
+
+    // Basic default prover config using the session_id returned from /session endpoint just now
     let config = ProverConfig::builder()
         .id(session_id)
         .server_dns(SERVER_DOMAIN)
-        .max_transcript_size(MAX_TRANSCRIPT_SIZE)
         .build()
         .unwrap();
 
-    // Create a Prover and set it up with the Notary
-    // This will set up the MPC backend prior to connecting to the server.
+    // Create a new prover and set up the MPC backend.
     let prover = Prover::new(config)
-        .setup(notary_socket.compat())
+        .setup(notary_tls_socket.compat())
         .await
         .unwrap();
 
-    // Connect to the Server via TCP. This is the TLS client socket.
     let client_socket = tokio::net::TcpStream::connect((SERVER_DOMAIN, 443))
         .await
         .unwrap();
 
-    // Bind the Prover to the server connection.
-    // The returned `mpc_tls_connection` is an MPC TLS connection to the Server: all data written
-    // to/read from it will be encrypted/decrypted using MPC with the Notary.
-    let (mpc_tls_connection, prover_fut) = prover.connect(client_socket.compat()).await.unwrap();
+    // Bind the Prover to server connection
+    let (tls_connection, prover_fut) = prover.connect(client_socket.compat()).await.unwrap();
 
-    // Spawn the Prover task to be run concurrently
+    // Spawn the Prover to be run concurrently
     let prover_task = tokio::spawn(prover_fut);
 
-    // Attach the hyper HTTP client to the MPC TLS connection
-    let (mut request_sender, connection) =
-        hyper::client::conn::handshake(mpc_tls_connection.compat())
-            .await
-            .unwrap();
+    // Attach the hyper HTTP client to the TLS connection
+    let (mut request_sender, connection) = hyper::client::conn::handshake(tls_connection.compat())
+        .await
+        .unwrap();
 
     // Spawn the HTTP task to be run concurrently
     let connection_task = tokio::spawn(connection.without_shutdown());
 
-    // Build a simple HTTP request with common headers
+    // Build the HTTP request to fetch the DMs
     let request = Request::builder()
-        .uri("/")
+        .uri(format!(
+            "https://{SERVER_DOMAIN}/{ROUTE}/{conversation_id}.json"
+        ))
         .header("Host", SERVER_DOMAIN)
         .header("Accept", "*/*")
-        // Using "identity" instructs the Server not to use compression for its HTTP response.
-        // TLSNotary tooling does not support compression.
         .header("Accept-Encoding", "identity")
         .header("Connection", "close")
         .header("User-Agent", USER_AGENT)
+        .header("Authorization", format!("Bearer {access_token}"))
+        .header(
+            "Cookie",
+            format!("auth_token={auth_token}; ct0={csrf_token}"),
+        )
+        .header("Authority", SERVER_DOMAIN)
+        .header("X-Twitter-Auth-Type", "OAuth2Session")
+        .header("x-twitter-active-user", "yes")
+        .header("X-Client-Uuid", client_uuid)
+        .header("X-Csrf-Token", csrf_token.clone())
         .body(Body::empty())
         .unwrap();
 
-    println!("Starting an MPC TLS connection with the server");
+    debug!("Sending request");
 
-    // Send the request to the Server and get a response via the MPC TLS connection
     let response = request_sender.send_request(request).await.unwrap();
 
-    println!("Got a response from the server");
+    debug!("Sent request");
 
     assert!(response.status() == StatusCode::OK);
+
+    debug!("Request OK");
+
+    // Pretty printing :)
+    let payload = to_bytes(response.into_body()).await.unwrap().to_vec();
+    let parsed =
+        serde_json::from_str::<serde_json::Value>(&String::from_utf8_lossy(&payload)).unwrap();
+    debug!("{}", serde_json::to_string_pretty(&parsed).unwrap());
 
     // Close the connection to the server
     let mut client_socket = connection_task.await.unwrap().unwrap().io.into_inner();
     client_socket.close().await.unwrap();
 
-    // The Prover task should be done now, so we can grab the Prover.
+    // The Prover task should be done now, so we can grab it.
     let prover = prover_task.await.unwrap().unwrap();
 
-    // Prepare for notarization.
+    // Prepare for notarization
     let mut prover = prover.start_notarize();
 
-    // Identify the ranges in the outbound data which contain data which we want to disclose
-    let (public_ranges, _) = find_ranges(
+    // Identify the ranges in the transcript that contain secrets
+    let (public_ranges, private_ranges) = find_ranges(
         prover.sent_transcript().data(),
         &[
-            // Redact the value of the "User-Agent" header. It will NOT be disclosed.
-            USER_AGENT.as_bytes(),
+            access_token.as_bytes(),
+            auth_token.as_bytes(),
+            csrf_token.as_bytes(),
         ],
     );
 
@@ -122,52 +167,32 @@ async fn main() {
 
     let builder = prover.commitment_builder();
 
-    // Commit to each range of the public outbound data which we want to disclose
-    let sent_commitments: Vec<_> = public_ranges
-        .iter()
-        .map(|r| builder.commit_sent(r.clone()).unwrap())
-        .collect();
+    // Commit to the outbound transcript, isolating the data that contain secrets
+    for range in public_ranges.iter().chain(private_ranges.iter()) {
+        builder.commit_sent(range.clone()).unwrap();
+    }
 
-    // Commit to all inbound data in one shot, as we don't need to redact anything in it
-    let recv_commitment = builder.commit_recv(0..recv_len).unwrap();
+    // Commit to the full received transcript in one shot, as we don't need to redact anything
+    builder.commit_recv(0..recv_len).unwrap();
 
     // Finalize, returning the notarized session
     let notarized_session = prover.finalize().await.unwrap();
 
-    // Create a proof for all committed data in this session
-    let mut proof_builder = notarized_session.data().build_substrings_proof();
+    debug!("Notarization complete!");
 
-    // Reveal all the public ranges
-    for commitment_id in sent_commitments {
-        proof_builder.reveal(commitment_id).unwrap();
-    }
-    proof_builder.reveal(recv_commitment).unwrap();
-
-    let substrings_proof = proof_builder.build().unwrap();
-
-    let proof = TlsProof {
-        session: notarized_session.session_proof(),
-        substrings: substrings_proof,
-    };
-
-    // Write the proof to a file
-    let mut file = tokio::fs::File::create("proof.json").await.unwrap();
-    file.write_all(serde_json::to_string_pretty(&proof).unwrap().as_bytes())
-        .await
-        .unwrap();
-
-    println!("Notarization completed successfully!");
-    println!("The proof has been written to proof.json");
+    // Dump the notarized session to a file
+    let mut file = tokio::fs::File::create("twitter_dm.json").await.unwrap();
+    file.write_all(
+        serde_json::to_string_pretty(&notarized_session)
+            .unwrap()
+            .as_bytes(),
+    )
+    .await
+    .unwrap();
 }
 
-/// Connects to the Notary and sets up a notarization session.
-///
-/// Returns a socket used to communicate with the Notary and a notarization session id.
-async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
-    // Connect to the Notary via TLS
-
-    // Since the Notary will present a self-signed certificate, we add the CA which signed the
-    // certificate to the trusted list
+async fn setup_notary_connection() -> (tokio_rustls::client::TlsStream<TcpStream>, String) {
+    // Connect to the Notary via TLS-TCP
     let mut certificate_file_reader = read_pem_file(NOTARY_CA_CERT_PATH).await.unwrap();
     let mut certificates: Vec<Certificate> = rustls_pemfile::certs(&mut certificate_file_reader)
         .unwrap()
@@ -179,13 +204,12 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
     let mut root_store = RootCertStore::empty();
     root_store.add(&certificate).unwrap();
 
-    let client_config = ClientConfig::builder()
+    let client_notary_config = ClientConfig::builder()
         .with_safe_defaults()
         .with_root_certificates(root_store)
         .with_no_client_auth();
-    let notary_connector = TlsConnector::from(Arc::new(client_config));
+    let notary_connector = TlsConnector::from(Arc::new(client_notary_config));
 
-    // Establish a TCP connection to the notary
     let notary_socket = tokio::net::TcpStream::connect(SocketAddr::new(
         IpAddr::V4(NOTARY_HOST.parse().unwrap()),
         NOTARY_PORT,
@@ -193,14 +217,13 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
     .await
     .unwrap();
 
-    // Wrap the TCP connection in TLS
-    // Tell the TLS backend to expect that the Notary's cert was issued to "tlsnotaryserver.io"
     let notary_tls_socket = notary_connector
+        // Require the domain name of notary server to be the same as that in the server cert
         .connect("tlsnotaryserver.io".try_into().unwrap(), notary_socket)
         .await
         .unwrap();
 
-    // Attach the hyper HTTP client to the notary TLS connection and start the TLS connection
+    // Attach the hyper HTTP client to the notary TLS connection to send request to the /session endpoint to configure notarization and obtain session id
     let (mut request_sender, connection) = hyper::client::conn::handshake(notary_tls_socket)
         .await
         .unwrap();
@@ -211,9 +234,10 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
     // Build the HTTP request to configure notarization
     let payload = serde_json::to_string(&NotarizationSessionRequest {
         client_type: ClientType::Tcp,
-        max_transcript_size: Some(MAX_TRANSCRIPT_SIZE),
+        max_transcript_size: Some(NOTARY_MAX_TRANSCRIPT_SIZE),
     })
     .unwrap();
+
     let request = Request::builder()
         .uri(format!("https://{NOTARY_HOST}:{NOTARY_PORT}/session"))
         .method("POST")
@@ -234,19 +258,17 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
     debug!("Response OK");
 
     // Pretty printing :)
-    let configuration_response = to_bytes(configuration_response.into_body())
+    let payload = to_bytes(configuration_response.into_body())
         .await
         .unwrap()
         .to_vec();
-    let configuration_response = serde_json::from_str::<NotarizationSessionResponse>(
-        &String::from_utf8_lossy(&configuration_response),
-    )
-    .unwrap();
+    let notarization_response =
+        serde_json::from_str::<NotarizationSessionResponse>(&String::from_utf8_lossy(&payload))
+            .unwrap();
 
-    debug!("Configuration response: {:?}", configuration_response,);
+    debug!("Notarization response: {:?}", notarization_response,);
 
-    // Request the notary to prepare for notarization via HTTP, where the underlying TCP connection
-    // will be extracted later
+    // Send notarization request via HTTP, where the underlying TCP connection will be extracted later
     let request = Request::builder()
         // Need to specify the session_id so that notary server knows the right configuration to use
         // as the configuration is set in the previous HTTP call
@@ -254,7 +276,7 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
             "https://{}:{}/notarize?sessionId={}",
             NOTARY_HOST,
             NOTARY_PORT,
-            configuration_response.session_id.clone()
+            notarization_response.session_id.clone()
         ))
         .method("GET")
         .header("Host", NOTARY_HOST)
@@ -264,11 +286,11 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
         .body(Body::empty())
         .unwrap();
 
-    debug!("Sending notarization preparation request");
+    debug!("Sending notarization request");
 
     let response = request_sender.send_request(request).await.unwrap();
 
-    debug!("Sent notarization preparation request");
+    debug!("Sent notarization request");
 
     assert!(response.status() == StatusCode::SWITCHING_PROTOCOLS);
 
@@ -280,15 +302,15 @@ async fn connect_to_notary() -> (TlsStream<TcpStream>, String) {
         ..
     } = connection_task.await.unwrap().unwrap();
 
-    (notary_tls_socket, configuration_response.session_id)
+    (notary_tls_socket, notarization_response.session_id)
 }
 
 /// Find the ranges of the public and private parts of a sequence.
 ///
 /// Returns a tuple of `(public, private)` ranges.
-fn find_ranges(seq: &[u8], private_seq: &[&[u8]]) -> (Vec<Range<usize>>, Vec<Range<usize>>) {
+fn find_ranges(seq: &[u8], sub_seq: &[&[u8]]) -> (Vec<Range<usize>>, Vec<Range<usize>>) {
     let mut private_ranges = Vec::new();
-    for s in private_seq {
+    for s in sub_seq {
         for (idx, w) in seq.windows(s.len()).enumerate() {
             if w == *s {
                 private_ranges.push(idx..(idx + w.len()));
